chore: refactor gen/docs.go
Split format-specific generation implementations into separate files
- markdown --> docs_md.go
- html-page --> docs_html.go
- html-embed --> docs_html.go
